feat(auth): extracts auth components to separate folder
