+++ components/Signup/Header.tsx
@@ -1,0 +1,12 @@
+import Head from "next/head";
+
+const Header = (props) => {
+    const { title } = props;
+    return (
+        <Head>
+            <title>{title}</title>
+        </Head>
+    );
+};
+
+export default Header;
